Use pkg-config to find include directories when on linux
Libdvidcpp does not compile against the system libraries on Debian and Ubuntu. Both of these distributions place the header files under /usr/include/jsonccp/json/ instead of /usr/include/json/. The compilation of libdvidcpp then fails due to missing header files.

Luckily, libjsoncpp ships with a .pc file for pkg-config. With the changes in this commit, cmake calls to pkg-config in order to query the proper include directory. Because OS X and Windows do not ship with pkg-config installed by default, this change only applies to Linux for now.